+++ src/test/java/com/bitbank/controllers/PingControllerTest.java
@@ -1,0 +1,60 @@
+package com.bitbank.controllers;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.web.servlet.MockMvc;
+
+import com.bitbank.config.AuthEntryPointJwt;
+import com.bitbank.config.AuthTokenFilter;
+import com.bitbank.controllers.PingController;
+import com.bitbank.services.PingService;
+import com.bitbank.services.UserDetailsServiceImpl;
+import com.bitbank.utils.JwtUtils;
+
+@ExtendWith(SpringExtension.class)
+@WebMvcTest(PingController.class)
+@TestPropertySource("classpath:application.properties")
+class PingControllerTest {
+
+	@Autowired
+	private MockMvc mvc;
+
+	@Autowired
+	AuthTokenFilter authTokenFilter;
+
+	@MockBean
+	AuthEntryPointJwt authEntryPointJwt;
+
+	@MockBean
+	JwtUtils jwtUtils;
+
+	@MockBean
+	UserDetailsServiceImpl userDetailsServiceImpl;
+
+	@MockBean
+	private PingService pingService;
+
+	@Value("${app.version}")
+	private String appVersion;
+
+	@Test
+	void indexShouldReturnApiVersion() throws Exception {
+
+		when(pingService.show()).thenReturn(appVersion);
+
+		mvc.perform(get("/api/v1/ping")).andExpect(status().isOk()).andExpect(jsonPath("$.version", is(appVersion)));
+	}
+
+}